Codify OW patch
