+++ adapters/pubmatic/pubmatic.go
@@ -536,44 +536,3 @@ func Builder(bidderName openrtb_ext.BidderName, config config.Adapter) (adapters
 	}
 	return bidder, nil
 }
-
-func getTargetingKeys(bidExt json.RawMessage, bidderName string) map[string]string {
-	targets := map[string]string{}
-	if bidExt != nil {
-		bidExtMap := make(map[string]interface{})
-		err := json.Unmarshal(bidExt, &bidExtMap)
-		if err == nil && bidExtMap[buyId] != nil {
-			targets[buyIdTargetingKey+bidderName] = string(bidExtMap[buyId].(string))
-		}
-	}
-	return targets
-}
-
-func copySBExtToBidExt(sbExt json.RawMessage, bidExt json.RawMessage) json.RawMessage {
-	if sbExt != nil {
-		sbExtMap := getMapFromJSON(sbExt)
-		bidExtMap := make(map[string]interface{})
-		if bidExt != nil {
-			bidExtMap = getMapFromJSON(bidExt)
-		}
-		if bidExtMap != nil && sbExtMap != nil {
-			if sbExtMap[buyId] != nil && bidExtMap[buyId] == nil {
-				bidExtMap[buyId] = sbExtMap[buyId]
-			}
-		}
-		byteAra, _ := json.Marshal(bidExtMap)
-		return json.RawMessage(byteAra)
-	}
-	return bidExt
-}
-
-func getMapFromJSON(ext json.RawMessage) map[string]interface{} {
-	if ext != nil {
-		extMap := make(map[string]interface{})
-		err := json.Unmarshal(ext, &extMap)
-		if err == nil {
-			return extMap
-		}
-	}
-	return nil
-}
