+++ adapters/pubmatic/pubmatic_test.go
@@ -71,88 +71,6 @@ func TestGetBidTypeForUnsupportedCode(t *testing.T) {
 	}
 }
 
-func TestGetAdServerTargetingForEmptyExt(t *testing.T) {
-	ext := json.RawMessage(`{}`)
-	targets := getTargetingKeys(ext, "pubmatic")
-	// banner is the default bid type when no bidType key is present in the bid.ext
-	if targets != nil && targets["hb_buyid_pubmatic"] != "" {
-		t.Errorf("It should not contained AdserverTageting")
-	}
-}
-
-func TestGetAdServerTargetingForValidExt(t *testing.T) {
-	ext := json.RawMessage("{\"buyid\":\"testBuyId\"}")
-	targets := getTargetingKeys(ext, "pubmatic")
-	// banner is the default bid type when no bidType key is present in the bid.ext
-	if targets == nil {
-		t.Error("It should have targets")
-		t.FailNow()
-	}
-	if targets != nil && targets["hb_buyid_pubmatic"] != "testBuyId" {
-		t.Error("It should have testBuyId as targeting")
-		t.FailNow()
-	}
-}
-
-func TestGetAdServerTargetingForPubmaticAlias(t *testing.T) {
-	ext := json.RawMessage("{\"buyid\":\"testBuyId-alias\"}")
-	targets := getTargetingKeys(ext, "dummy-alias")
-	// banner is the default bid type when no bidType key is present in the bid.ext
-	if targets == nil {
-		t.Error("It should have targets")
-		t.FailNow()
-	}
-	if targets != nil && targets["hb_buyid_dummy-alias"] != "testBuyId-alias" {
-		t.Error("It should have testBuyId as targeting")
-		t.FailNow()
-	}
-}
-
-func TestGetMapFromJSON(t *testing.T) {
-	ext := json.RawMessage("{\"buyid\":\"testBuyId\"}")
-	extMap := getMapFromJSON(ext)
-	if extMap == nil {
-		t.Errorf("it should be converted in extMap")
-	}
-}
-
-func TestGetMapFromJSONWithInvalidJSON(t *testing.T) {
-	ext := json.RawMessage("{\"buyid\":\"testBuyId\"}}}}")
-	extMap := getMapFromJSON(ext)
-	if extMap != nil {
-		t.Errorf("it should be converted in extMap")
-	}
-}
-
-func TestCopySBExtToBidExtWithBidExt(t *testing.T) {
-	sbext := json.RawMessage("{\"buyid\":\"testBuyId\"}")
-	bidext := json.RawMessage("{\"dspId\":\"9\"}")
-	// expectedbid := json.RawMessage("{\"dspId\":\"9\",\"buyid\":\"testBuyId\"}")
-	bidextnew := copySBExtToBidExt(sbext, bidext)
-	if bidextnew == nil {
-		t.Errorf("it should not be nil")
-	}
-}
-
-func TestCopySBExtToBidExtWithNoBidExt(t *testing.T) {
-	sbext := json.RawMessage("{\"buyid\":\"testBuyId\"}")
-	bidext := json.RawMessage("{\"dspId\":\"9\"}")
-	// expectedbid := json.RawMessage("{\"dspId\":\"9\",\"buyid\":\"testBuyId\"}")
-	bidextnew := copySBExtToBidExt(sbext, bidext)
-	if bidextnew == nil {
-		t.Errorf("it should not be nil")
-	}
-}
-
-func TestCopySBExtToBidExtWithNoSeatExt(t *testing.T) {
-	bidext := json.RawMessage("{\"dspId\":\"9\"}")
-	// expectedbid := json.RawMessage("{\"dspId\":\"9\",\"buyid\":\"testBuyId\"}")
-	bidextnew := copySBExtToBidExt(nil, bidext)
-	if bidextnew == nil {
-		t.Errorf("it should not be nil")
-	}
-}
-
 func TestParseImpressionObject(t *testing.T) {
 	type args struct {
 		imp                      *openrtb2.Imp
