+++ router/router.go
@@ -2,63 +2,34 @@ package router
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/prebid/prebid-server/analytics"
-	"github.com/prebid/prebid-server/stored_requests"
-	"github.com/prebid/prebid-server/usersync"
-	"github.com/prometheus/client_golang/prometheus"
-
 	analyticsConf "github.com/prebid/prebid-server/analytics/config"
+
+	"github.com/prebid/prebid-server/usersync"
+
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/currency"
-	"github.com/prebid/prebid-server/endpoints"
-	"github.com/prebid/prebid-server/endpoints/openrtb2"
 	"github.com/prebid/prebid-server/errortypes"
 	"github.com/prebid/prebid-server/exchange"
 	"github.com/prebid/prebid-server/gdpr"
-	"github.com/prebid/prebid-server/metrics"
 	metricsConf "github.com/prebid/prebid-server/metrics/config"
 	"github.com/prebid/prebid-server/openrtb_ext"
 	pbc "github.com/prebid/prebid-server/prebid_cache_client"
 	"github.com/prebid/prebid-server/server/ssl"
 	storedRequestsConf "github.com/prebid/prebid-server/stored_requests/config"
 	"github.com/prebid/prebid-server/util/sliceutil"
-	"github.com/prebid/prebid-server/util/uuidutil"
 
 	"github.com/golang/glog"
 	"github.com/julienschmidt/httprouter"
 	_ "github.com/lib/pq"
 	"github.com/rs/cors"
-)
-
-var (
-	g_syncers           map[string]usersync.Syncer
-	g_cfg               *config.Configuration
-	g_ex                exchange.Exchange
-	g_accounts          stored_requests.AccountFetcher
-	g_paramsValidator   openrtb_ext.BidderParamValidator
-	g_storedReqFetcher  stored_requests.Fetcher
-	g_storedRespFetcher stored_requests.Fetcher
-	g_gdprPerms         gdpr.Permissions
-	g_metrics           metrics.MetricsEngine
-	g_analytics         analytics.PBSAnalyticsModule
-	g_disabledBidders   map[string]string
-	g_categoriesFetcher stored_requests.CategoryFetcher
-	g_videoFetcher      stored_requests.Fetcher
-	g_activeBidders     map[string]openrtb_ext.BidderName
-	g_defReqJSON        []byte
-	g_cacheClient       pbc.Client
-	g_transport         *http.Transport
 )
 
 // NewJsonDirectoryServer is used to serve .json files from a directory as a single blob. For example,
@@ -200,46 +171,43 @@ func New(cfg *config.Configuration, rateConvertor *currency.RateConverter) (r *R
 	}
 
 	var errs []error
-	g_syncers, errs = usersync.BuildSyncers(cfg, bidderInfos)
+	syncersByBidder, errs := usersync.BuildSyncers(cfg, bidderInfos)
 	if len(errs) > 0 {
 		return nil, errortypes.NewAggregateError("user sync", errs)
 	}
 
-	syncerKeys := make([]string, 0, len(g_syncers))
+	syncerKeys := make([]string, 0, len(syncersByBidder))
 	syncerKeysHashSet := map[string]struct{}{}
-	for _, syncer := range g_syncers {
+	for _, syncer := range syncersByBidder {
 		syncerKeysHashSet[syncer.Key()] = struct{}{}
 	}
 	for k := range syncerKeysHashSet {
 		syncerKeys = append(syncerKeys, k)
 	}
 
-	g_cfg = cfg
 	// Metrics engine
-	g_metrics = metricsConf.NewMetricsEngine(cfg, openrtb_ext.CoreBidderNames(), syncerKeys)
-	_, g_storedReqFetcher, _, g_accounts, g_categoriesFetcher, g_videoFetcher, g_storedRespFetcher = storedRequestsConf.NewStoredRequests(cfg, g_metrics, generalHttpClient, r.Router)
+	r.MetricsEngine = metricsConf.NewMetricsEngine(cfg, openrtb_ext.CoreBidderNames(), syncerKeys)
+	_, fetcher, _, accounts, categoriesFetcher, videoFetcher, storedRespFetcher := storedRequestsConf.NewStoredRequests(cfg, r.MetricsEngine, generalHttpClient, r.Router)
 	// todo(zachbadgett): better shutdown
 	// r.Shutdown = shutdown
 
-	g_analytics = analyticsConf.NewPBSAnalytics(&cfg.Analytics)
+	pbsAnalytics := analyticsConf.NewPBSAnalytics(&cfg.Analytics)
 
-	g_paramsValidator, err = openrtb_ext.NewBidderParamsValidator(schemaDirectory)
+	paramsValidator, err := openrtb_ext.NewBidderParamsValidator(schemaDirectory)
 	if err != nil {
 		glog.Fatalf("Failed to create the bidder params validator. %v", err)
 	}
 
-	g_activeBidders = exchange.GetActiveBidders(bidderInfos)
-	g_disabledBidders = exchange.GetDisabledBiddersErrorMessages(bidderInfos)
+	activeBidders := exchange.GetActiveBidders(bidderInfos)
+	disabledBidders := exchange.GetDisabledBiddersErrorMessages(bidderInfos)
 
 	defaultAliases, defReqJSON := readDefaultRequest(cfg.DefReqConfig)
 	if err := validateDefaultAliases(defaultAliases); err != nil {
 		return nil, err
 	}
 
-	g_defReqJSON = defReqJSON
-
 	gvlVendorIDs := bidderInfos.ToGVLVendorIDMap()
-	g_gdprPerms = gdpr.NewPermissions(context.Background(), cfg.GDPR, gvlVendorIDs, generalHttpClient)
+	gdprPerms := gdpr.NewPermissions(context.Background(), cfg.GDPR, gvlVendorIDs, generalHttpClient)
 
 	if cfg.VendorListScheduler.Enabled {
 		vendorListScheduler, err := gdpr.GetVendorListScheduler(cfg.VendorListScheduler.Interval, cfg.VendorListScheduler.Timeout, generalHttpClient)
@@ -249,15 +217,15 @@ func New(cfg *config.Configuration, rateConvertor *currency.RateConverter) (r *R
 		vendorListScheduler.Start()
 	}
 
-	g_cacheClient = pbc.NewClient(cacheHttpClient, &cfg.CacheURL, &cfg.ExtCacheURL, g_metrics)
+	cacheClient := pbc.NewClient(cacheHttpClient, &cfg.CacheURL, &cfg.ExtCacheURL, r.MetricsEngine)
 
-	adapters, adaptersErrs := exchange.BuildAdapters(generalHttpClient, cfg, bidderInfos, g_metrics)
+	adapters, adaptersErrs := exchange.BuildAdapters(generalHttpClient, cfg, bidderInfos, r.MetricsEngine)
 	if len(adaptersErrs) > 0 {
 		errs := errortypes.NewAggregateError("Failed to initialize adapters", adaptersErrs)
 		return nil, errs
 	}
 
-	g_ex = exchange.NewExchange(adapters, g_cacheClient, cfg, g_syncers, g_metrics, bidderInfos, g_gdprPerms, rateConvertor, g_categoriesFetcher)
+	theExchange := exchange.NewExchange(adapters, cacheClient, cfg, syncersByBidder, r.MetricsEngine, bidderInfos, gdprPerms, rateConvertor, categoriesFetcher)
 	/*var uuidGenerator uuidutil.UUIDRandomGenerator
 	openrtbEndpoint, err := openrtb2.NewEndpoint(uuidGenerator, theExchange, paramsValidator, fetcher, accounts, cfg, r.MetricsEngine, pbsAnalytics, disabledBidders, defReqJSON, activeBidders, storedRespFetcher)
 	if err != nil {
@@ -312,40 +280,24 @@ func New(cfg *config.Configuration, rateConvertor *currency.RateConverter) (r *R
 	r.POST("/optout", userSyncDeps.OptOut)
 	r.GET("/optout", userSyncDeps.OptOut)*/
 
+	g_syncers = syncersByBidder
+	g_metrics = r.MetricsEngine
+	g_cfg = cfg
+	g_storedReqFetcher = &fetcher
+	g_accounts = &accounts
+	g_categoriesFetcher = &categoriesFetcher
+	g_videoFetcher = &videoFetcher
+	g_storedRespFetcher = &storedRespFetcher
+	g_analytics = &pbsAnalytics
+	g_paramsValidator = &paramsValidator
+	g_activeBidders = activeBidders
+	g_disabledBidders = disabledBidders
+	g_defReqJSON = defReqJSON
+	g_gdprPerms = &gdprPerms
+	g_cacheClient = &cacheClient
+	g_ex = &theExchange
+
 	return r, nil
-}
-
-func getTransport(cfg *config.Configuration, certPool *x509.CertPool) *http.Transport {
-	transport := &http.Transport{
-		MaxConnsPerHost: cfg.Client.MaxConnsPerHost,
-		IdleConnTimeout: time.Duration(cfg.Client.IdleConnTimeout) * time.Second,
-		TLSClientConfig: &tls.Config{RootCAs: certPool},
-	}
-
-	if cfg.Client.DialTimeout > 0 {
-		transport.Dial = (&net.Dialer{
-			Timeout:   time.Duration(cfg.Client.DialTimeout) * time.Millisecond,
-			KeepAlive: time.Duration(cfg.Client.DialKeepAlive) * time.Second,
-		}).Dial
-	}
-
-	if cfg.Client.TLSHandshakeTimeout > 0 {
-		transport.TLSHandshakeTimeout = time.Duration(cfg.Client.TLSHandshakeTimeout) * time.Second
-	}
-
-	if cfg.Client.ResponseHeaderTimeout > 0 {
-		transport.ResponseHeaderTimeout = time.Duration(cfg.Client.ResponseHeaderTimeout) * time.Second
-	}
-
-	if cfg.Client.MaxIdleConns > 0 {
-		transport.MaxIdleConns = cfg.Client.MaxIdleConns
-	}
-
-	if cfg.Client.MaxIdleConnsPerHost > 0 {
-		transport.MaxIdleConnsPerHost = cfg.Client.MaxIdleConnsPerHost
-	}
-
-	return transport
 }
 
 func applyBidderInfoConfigOverrides(bidderInfos config.BidderInfos, adaptersCfg map[string]config.Adapter) error {
@@ -429,66 +381,6 @@ func checkSupportedUserSyncEndpoints(bidderInfos config.BidderInfos) error {
 		}
 	}
 	return nil
-}
-
-func GetCacheClient() *pbc.Client {
-	return &g_cacheClient
-}
-
-func GetPrebidCacheURL() string {
-	return g_cfg.ExternalURL
-}
-
-//OrtbAuctionEndpointWrapper Openwrap wrapper method for calling /openrtb2/auction endpoint
-func OrtbAuctionEndpointWrapper(w http.ResponseWriter, r *http.Request) error {
-	ortbAuctionEndpoint, err := openrtb2.NewEndpoint(uuidutil.UUIDRandomGenerator{}, g_ex, g_paramsValidator, g_storedReqFetcher, g_accounts, g_cfg, g_metrics, g_analytics, g_disabledBidders, g_defReqJSON, g_activeBidders, g_storedRespFetcher)
-	if err != nil {
-		return err
-	}
-	ortbAuctionEndpoint(w, r, nil)
-	return nil
-}
-
-//VideoAuctionEndpointWrapper Openwrap wrapper method for calling /openrtb2/video endpoint
-func VideoAuctionEndpointWrapper(w http.ResponseWriter, r *http.Request) error {
-	videoAuctionEndpoint, err := openrtb2.NewCTVEndpoint(g_ex, g_paramsValidator, g_storedReqFetcher, g_videoFetcher, g_accounts, g_cfg, g_metrics, g_analytics, g_disabledBidders, g_defReqJSON, g_activeBidders)
-	if err != nil {
-		return err
-	}
-	videoAuctionEndpoint(w, r, nil)
-	return nil
-}
-
-//GetUIDSWrapper Openwrap wrapper method for calling /getuids endpoint
-func GetUIDSWrapper(w http.ResponseWriter, r *http.Request) {
-	getUID := endpoints.NewGetUIDsEndpoint(g_cfg.HostCookie)
-	getUID(w, r, nil)
-}
-
-//SetUIDSWrapper Openwrap wrapper method for calling /setuid endpoint
-func SetUIDSWrapper(w http.ResponseWriter, r *http.Request) {
-	setUID := endpoints.NewSetUIDEndpoint(g_cfg.HostCookie, g_syncers, g_gdprPerms, g_analytics, g_metrics)
-	setUID(w, r, nil)
-}
-
-//CookieSync Openwrap wrapper method for calling /cookie_sync endpoint
-func CookieSync(w http.ResponseWriter, r *http.Request) {
-	cookiesync := endpoints.NewCookieSyncEndpoint(g_syncers, g_cfg, g_gdprPerms, g_metrics, g_analytics, g_activeBidders)
-	cookiesync.Handle(w, r, nil)
-}
-
-//SyncerMap Returns map of bidder and its usersync info
-func SyncerMap() map[string]usersync.Syncer {
-	return g_syncers
-}
-
-func GetPrometheusGatherer() *prometheus.Registry {
-	mEngine, ok := g_metrics.(*metricsConf.DetailedMetricsEngine)
-	if !ok || mEngine == nil || mEngine.PrometheusMetrics == nil {
-		return nil
-	}
-
-	return mEngine.PrometheusMetrics.Gatherer
 }
 
 // Fixes #648
